fixing test by removing unneccessary gate operations
in a seperate pull request to the pyquil repo we will
propose to remove this sequence of gates that implement
the identity operation. First of all, the operation is superfluous
and second of all it messes up the wavefunction calculation since
qubit 0 was hardcoded to always be used for this sequence of gates.
It becomes a problem if we embedded our problem on qubit 3,4 and want
to get the wavefunction since it will always try to involve qubit 0.